requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
click==8.1.7

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Configuration and Utilities
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Async Support
asyncio-mqtt==0.16.1

# HTTP Client
httpx==0.25.2

# File handling
pathlib2==2.3.7

# Date/Time handling
python-dateutil==2.8.2 